+++ test/fixtures/noLLMToolAssistant.js
@@ -3,10 +3,10 @@ import { Assistant, Tool } from "../../src/index.js";
 
 class AnswerTool extends Tool {
   constructor() {
-    const name = helperName("NoLLMAnswerTool");
-    const description = "Answers to messages.";
-    const instructions = description;
-    super(name, description, instructions, {
+    super({
+      name: helperName("NoLLMAnswerTool"),
+      description: "Answers to messages.",
+      instructions: "Answers to messages.",
       llm: false,
       temperature: 0.1,
       parentsTools: [
@@ -14,7 +14,7 @@ class AnswerTool extends Tool {
           type: "function",
           function: {
             name: AnswerTool.toolName,
-            description: description,
+            description: "Answers to messages.",
             parameters: {
               type: "object",
               properties: { message: { type: "string" } },
@@ -33,10 +33,11 @@ class AnswerTool extends Tool {
 
 class NoLLMToolAssistant extends Assistant {
   constructor() {
-    const name = helperName("NoLLMToolAssistant");
-    const description = "Answers to messages.";
-    const instructions = `You must route /tool messages in full to your '${AnswerTool.toolName}' tool. Never respond without first using that tool. Never! Ex: When asked what color the sky is, use the '${AnswerTool.toolName}' tool first.`;
-    super(name, description, instructions);
+    super({
+      name: helperName("NoLLMToolAssistant"),
+      description: "Answers to messages.",
+      instructions: `You must route /tool messages in full to your '${AnswerTool.toolName}' tool. Never respond without first using that tool. Never! Ex: When asked what color the sky is, use the '${AnswerTool.toolName}' tool first.`,
+    });
     this.addAssistantTool(AnswerTool);
   }
 }